@@ -1,6 +1,16 @@
 <template>
   <div class="home-wrap">
     Home
+    <div>
+      <input
+        v-model="username"
+        type="text"
+      >
+      <button @click="setUsername">
+        set username
+      </button>
+      <div>username: {{ this.$store.state.user.name }}</div>
+    </div>
     <h2>home router-view</h2>
     <router-view />
   </div>
@@ -9,7 +19,15 @@
 <script>
 
 export default {
-  components: {
+  data() {
+    return {
+      username: '',
+    };
+  },
+  methods: {
+    setUsername() {
+      this.$store.commit('setUser', { name: this.username });
+    },
   },
 };
 </script>
